Revise test cases completely
Due to the revision of the DeadCodeChangeAnalyzer and, in particular,
the identification of relevant changes to code blocks, we can simplify
the test cases. This is, we only need to check for code files, in which
block statements are changed. There is no differentiation between all
blocks and blocks referencing configuration options as the analyzer does
not offer this option anymore.